FUCK MY PRODUCT MANAGER! istg
We had a *perfectly fine* preferential voting system. It wasn't perfect - of course. No voting system is. But it was perfectly *fine*. Everone else (ie. one guy - apart from the said annoying PM) on the election commiete agreed. Just a good'ol score for every vote. Like a lot of them work. It was OK.

Then, this mf comes along, and says that scoring linearly is a bad idea; „undemocratic”. The „one guy” mentioned earlier switched sides, and says the same. But has an opposite idea! PM prefers 1,2,3,5,7 (increase the weight of higher options), some guy prefers 1,1,2,2,3 (decrease them). Why none of them support the linear system? I DUNNO! It's the easiest.

We ended up arguing for way too long about it  - and setteled on..... 1 1 2 3 4. YES, IT IS almost the same. Just... One lower, and options 5 and 4 have the same weigth. You think it's dumb?

So do I! But it's supposed to somehow be a compromise between my suggestion (linear - cuz it's almost the same as it; just one point lower and 0+voted-on-this=1), PM's (becasue it „synergistically highlights the strategic importance of votes 1 2 and 3, and their critical superiority over 4 and 5, as they're better than those with only 1 point” - ok, he didn't say that exaclty, as there were no corporate buzzwords, unfortunatley) and the other guy's (higher votes got devalued in comparison to lower).

why is my team like this

it's the second time we spent way too much time arguing over meaningless bs and picked the most awkward solution as a „compromise”